Rule logic for returning context around cloze
This implementation works by putting a marker on the revealed cloze tag,
casting the entire HTML soup and the revealed cloze tag into a string,
and then splitting on the string version of the revealed cloze tag.

Doing it this way feels messy, but it is by far the most reliable way I
found, as this allows me to retain spaces/absence-of-spaces adjacent to
the revealed cloze and also avoids all the madness with counting HTML
entities and traversing both `parentSibling` and `parent`.